+++ core/src/cn/harryh/arkpets/ArkPets.java
@@ -8,7 +8,7 @@ import cn.harryh.arkpets.animations.AnimData;
 import cn.harryh.arkpets.animations.GeneralBehavior;
 import cn.harryh.arkpets.concurrent.SocketClient;
 import cn.harryh.arkpets.platform.HWndCtrl;
-import cn.harryh.arkpets.platform.WaylandHWnd;
+import cn.harryh.arkpets.platform.WaylandHWndCtrl;
 import cn.harryh.arkpets.platform.WindowSystem;
 import cn.harryh.arkpets.render.PixmapWrapper;
 import cn.harryh.arkpets.transitions.TransitionVector2;
@@ -112,7 +112,7 @@ public class ArkPets extends InputApplicationAdaptor {
         if (config.window_style_toolwindow)
             hWndMine.setTaskbar(false);
         updateWindow();
-        if (hWndMine instanceof WaylandHWnd wlHWnd) wlHWnd.initSurface((Lwjgl3Graphics) Gdx.graphics);
+        if (hWndMine instanceof WaylandHWndCtrl wlHWnd) wlHWnd.attachSurface((Lwjgl3Graphics) Gdx.graphics);
 
         // 6.Tray icon setup
         tray = new MemberTrayImpl(this, new SocketClient());
