+++ core/src/cn/harryh/arkpets/platform/KWinHWndCtrl.java
@@ -3,9 +3,7 @@ package cn.harryh.arkpets.platform;
 import cn.harryh.arkpets.Const;
 import cn.harryh.arkpets.natives.KWinInterface;
 import cn.harryh.arkpets.natives.KWinPluginInterface;
-import cn.harryh.arkpets.natives.WaylandHelper;
 import cn.harryh.arkpets.utils.Logger;
-import com.sun.jna.Pointer;
 import org.freedesktop.dbus.connections.impl.DBusConnection;
 import org.freedesktop.dbus.connections.impl.DBusConnectionBuilder;
 import org.freedesktop.dbus.exceptions.DBusException;
@@ -17,14 +15,11 @@ import java.util.Collections;
 import java.util.List;
 
 
-public class KWinHWndCtrl extends HWndCtrl implements WaylandHWnd {
+public class KWinHWndCtrl extends WaylandHWndCtrl {
     protected final String hWnd;
     protected KWinInterface.DetailsStruct details;
     private static DBusConnection dBusConnection;
     private static KWinInterface dBusInterface;
-
-    private boolean isTransparent;
-    private Pointer surface;
 
     protected KWinHWndCtrl(KWinInterface.DetailsStruct details) {
         super(details.title, new WindowRect(details.y, details.y + details.h.intValue(), details.x, details.x + details.w.intValue()));
@@ -63,14 +58,6 @@ public class KWinHWndCtrl extends HWndCtrl implements WaylandHWnd {
     }
 
     @Override
-    public void setTransparent(boolean enable) {
-        if (isTransparent != enable) {
-            WaylandHelper.setTransparent(surface, enable);
-            isTransparent = enable;
-        }
-    }
-
-    @Override
     public void setTaskbar(boolean enable) {
         dBusInterface.Stick(hWnd, enable);
     }
@@ -83,10 +70,6 @@ public class KWinHWndCtrl extends HWndCtrl implements WaylandHWnd {
     @Override
     public void sendMouseEvent(MouseEvent msg, int x, int y) {
 
-    }
-
-    public void setSurface(Pointer surface) {
-        this.surface = surface;
     }
 
     protected static void init() {
